fix(forgejo): Patch to stop 500s on import+review
A while ago we set up forgejo. After that, we discovered that in certain
circumstances you could cause forgejo to give an error 500 when
approving a pull request.

Please see https://codeberg.org/forgejo/forgejo/issues/3860

This has been fixed upstream, but we are impatient so we would like to
pull in the patch before it lands in nixpkgs.